run deployments and some cluster setup tasks against first server only
This speeds things up as we now run a lot of steps which wouldn't
change the cluster anyway when executed on additional hosts only once,
but the implementation is not quite ideal:

1 it'll fail if the first server is unavailable
2 it'll not update deployments if the first server is not part of the
  host group this is run against

The simplicity by not adding logic for those is probably more
important than trying to automate those cornercases, though:

For 1 it only requires changing the first server IP in the
configuration.

For 2 not only would we need logic to select one node with server role
in the current play to take the role of temporary first server - we'd
also need to try figuring out if the first server was omitted
intentionally (and no updates to the deployment are expected). Just
requiring the first server to be present is significantly easier.